requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
livekit-api==0.4.3
requests==2.32.3
